PYIC-4119: Split Fraud tests into separate files
